Split BattleProcessor into few more parts
